Add the missing flag
As the maintainer of this package, I am an Egyptian and I personally support the Palestinian cause. However, I understand that the purpose of this package is to provide a collection of flag icons for use in software development, and it's not intended to take a stance on political issues. In the interest of being inclusive and respectful of all users.